requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
cfgv==3.4.0
colorama==0.4.6
distlib==0.3.7
distro==1.8.0
exceptiongroup==1.1.3
filelock==3.13.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
identify==2.5.31
idna==3.4
nodeenv==1.8.0
openai==1.2.3
platformdirs==3.11.0
pre-commit==3.5.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
virtualenv==20.24.6
